% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ricker.R
\name{get_ricker}
\alias{get_ricker}
\title{get_ricker returns the default parameters of the Ricker simulations}
\usage{
get_ricker()
}
\value{
data frame of default parameters for simulations
}
\description{
get_ricker returns the default parameters of the Ricker simulations
}
