% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_summary}
\alias{plot_summary}
\title{plot_summary returns basic plots of quantities of interest over time}
\usage{
plot_summary(summary_df)
}
\arguments{
\item{summary_df}{A dataframe generated from \code{summarize_simulations()}}
}
\value{
A list of plots
}
\description{
plot_summary returns basic plots of quantities of interest over time
}
