% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ricker.R
\name{ricker_defaults}
\alias{ricker_defaults}
\title{ricker_defaults returns the default parameters of the Ricker simulations}
\usage{
ricker_defaults()
}
\value{
data frame of default parameters for simulations
}
\description{
ricker_defaults returns the default parameters of the Ricker simulations
}
