% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{sim is the primary function for simulating management strategies to a regime switching model}
\usage{
sim(
  sims = 1000,
  time_steps = 100,
  ricker_pars = NULL,
  pr_12 = 0.1,
  pr_21 = 0.1,
  run = "odd",
  deterministic_model = TRUE,
  rec_std = 0.1,
  rec_acf = 0.7,
  escapement_rule = "both",
  harvest_CV = 0,
  discount_rate = 0.1,
  price_linear_change = 0,
  price_acf = 0.7,
  price_cv = 0,
  time_lag = 0,
  msy_scenario = c("equilibrium", "msy")[1],
  alpha = 0,
  seed = 123
)
}
\arguments{
\item{sims}{The number of simulations, defaults to 1000}

\item{time_steps}{The number of time steps, defaults to 100}

\item{ricker_pars}{a dataframe of custom parameters for the Ricker model}

\item{pr_12}{The probability of transitioning from state 1 to 2, defaults to 0.1}

\item{pr_21}{The probability of transitioning from state 1 to 2, defaults to 0.1}

\item{run}{Whether to run this for the odd or even broodline, defaults to "odd"}

\item{deterministic_model}{Whether the deterministic escapement rule is used, defaults to TRUE}

\item{rec_std}{Recruitment variability (lognormal sd), defaults to 0.1}

\item{rec_acf}{Temporal autocorrelation of recruitment. Values of 1 = random walk, values of 0 = white noise. Defaults to 0.7}

\item{escapement_rule}{Can be "pre", "post", or "both" (default). If "pre" or "post", a single
escapement rule is used}

\item{harvest_CV}{The coefficient of variation for harvest. Variability is lognormal -- and this parameter controls the amount of variation.}

\item{discount_rate}{Optional, can also be done after the fact, but defaults to 0.1}

\item{price_linear_change}{The price change over the entire time series. This is linear in log space to keep prices positive. Defaults to 0}

\item{price_acf}{The autocorrelation of prices. Defaults to 0.7, based on time series of pink salmon prices in Prince William Sound, 1984 - 2021}

\item{price_cv}{The coefficient of variation, or standard deviation in log space of prices. Defaults to 0 (realistic values based on time series of pink salmon prices in Prince William Sound, 1984 - 2021 = 0.1)}

\item{time_lag}{Optional, represents a delay in how fast response is to the actual regime. Defaults to
0 but could be a positive integer (e.g. 10) to represent delays in the decision making process}

\item{msy_scenario}{Scenario for maximum sustainable yield (MSY). Defaults to "equilibrium" where values are constant regardless of spawner abundance. Can
also be "msy", where values are updated each year}

\item{alpha}{The objective function to solve for with the MSY derivative. Defaults to 0}

\item{seed}{Seed for random number generation, defaults to 123}
}
\value{
data frame of simulations
}
\description{
sim is the primary function for simulating management strategies to a regime switching model
}
