% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{summarize_plots}
\alias{summarize_plots}
\title{summarize_plots returns basic plots of quantities of interest over time}
\usage{
summarize_plots(summary_df)
}
\arguments{
\item{summary_df}{A dataframe generated from \code{summarize_quantities()}}
}
\value{
A list of plots
}
\description{
summarize_plots returns basic plots of quantities of interest over time
}
