% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{fit_bycatch is the primary function for fitting bycatch models to time series of takes and effort}
\usage{
sim(
  sims = 1000,
  time_steps = 100,
  ricker_pars = NULL,
  pr_12 = 0.1,
  pr_21 = 0.1,
  run = "odd",
  deterministic_model = TRUE,
  rec_std = 0.1,
  rec_acf = 0.7,
  escapement_rule = "both",
  harvest_CV = 0,
  discount_rate = 0.1,
  time_lag = 0,
  seed = 123
)
}
\arguments{
\item{sims}{The number of simulations, defaults to 1000}

\item{time_steps}{The number of time steps, defaults to 100}

\item{ricker_pars}{a dataframe of custom parameters for the Ricker model}

\item{pr_12}{The probability of transitioning from state 1 to 2, defaults to 0.1}

\item{pr_21}{The probability of transitioning from state 1 to 2, defaults to 0.1}

\item{run}{Whether to run this for the odd or even broodline, defaults to "odd"}

\item{deterministic_model}{Whether the deterministic escapement rule is used, defaults to TRUE}

\item{rec_std}{Recruitment variability (lognormal sd), defaults to 0.1}

\item{rec_acf}{Temporal autocorrelation of recruitment. Values of 1 = random walk, values of 0 = white noise. Defaults to 0.7}

\item{escapement_rule}{Can be "pre", "post", or "both" (default). If "pre" or "post", a single
escapement rule is used}

\item{harvest_CV}{The coefficient of variation for harvest. Variability is lognormal -- and this parameter controls the amount of variation.}

\item{discount_rate}{Optional, can also be done after the fact, but defaults to 0.1}

\item{time_lag}{Optional, represents a delay in how fast response is to the actual regime. Defaults to
0 but could be a positive integer (e.g. 10) to represent delays in the decision making process}

\item{seed}{Seed for random number generation, defaults to 123}
}
\value{
data frame of simulations
}
\description{
fit_bycatch is the primary function for fitting bycatch models to time series of takes and effort
}
