% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_simulations.R
\name{summarize_simulations}
\alias{summarize_simulations}
\title{summarize_simulations returns summaries of benefits, harvest, and recruitment}
\usage{
summarize_simulations(df, quantiles = c(0.25, 0.75))
}
\arguments{
\item{df}{A dataframe. If 'scenario' is included as a variable, it is used for grouping}

\item{quantiles}{a 2 element vector containing the quantiles to be reported (along with mean and median). Defaults to c(0.25, 0.75)}
}
\value{
data frame of default parameters for simulations
}
\description{
summarize_simulations returns summaries of benefits, harvest, and recruitment
}
